* commit '0b69ad6a9ad8762eca044ee5faa098ab79638630 [formerly 8cd4b8c48e42f42b5b2c9c53f049324f0e27d35e]': (420 commits)
  DF-9998 fix testbed after merge development
  DF-9998 android testbed works, but without native controls
  DF-9998 rename Thread::GetCurrentIdAsInteger to GetCurrentIdAsUInt64
  DF-9998 android GLES: fix surface destruction
  DF-11294 fixed compilation, fixed ringbuffer size
  DF-185 fixed non-unity build compilation
  [DF-11072] added include <chrono>
  [DF-11285] Ignored Assert message at ErrorDialog
  [DF-10700] Fixed issue with reselection of entities and fixed scale for multiple objects
  [DF-11255] Fixed info about invisible grass
  DF-11188 Fix review issues
  [DF-10218] fix BOM
  [DF-11072] fix remark
  [DF-11072] reset ngt
  DF-11188 Fix review issues
  [DF-10700] Fixed applying of new values
  Revert "[DF-11072] reset commit"
  [DF-11072] reset commit
  DF-9998 android goes on
  [DF-11255] Removed empty dtor
  ...


Former-commit-id: fe6d53bda7e1879f486891517b881d3f56ba4bd4

+++ Projects/PerfomanceTests/Classes/Tests/BaseTest.h
@@ -108,7 +108,7 @@ protected:
     virtual void CreateUI();
     virtual void UpdateUI();
 
-    size_t GetAllocatedMemory();
+    uint32 GetAllocatedMemory();
 
     virtual void PerformTestLogic(float32 timeElapsed) = 0;
 
@@ -144,7 +144,7 @@ private:
     UIStaticText* elapsedTimeText;
     UIStaticText* framesRenderedText;
 
-    size_t maxAllocatedMemory;
+    uint32 maxAllocatedMemory;
 };
 
 inline const Vector<BaseTest::FrameInfo>& BaseTest::GetFramesInfo() const
